+++ dwds/test/fixtures/context.dart
@@ -180,7 +180,6 @@ class TestContext {
     _logger.info('Packages: $_packageConfigFile');
     _logger.info('Entry: $_dartEntryFilePath');
   }
-  static bool _sdkAssetsGenerated = false;
 
   Future<void> setUp({
     ReloadConfiguration reloadConfiguration = ReloadConfiguration.none,
@@ -196,7 +195,6 @@ class TestContext {
     CompilationMode compilationMode = CompilationMode.buildDaemon,
     bool enableExpressionEvaluation = false,
     bool verboseCompiler = false,
-    SdkConfigurationProvider? sdkConfigurationProvider,
     bool useDebuggerModuleNames = false,
     bool launchChrome = true,
     bool isFlutterApp = false,
@@ -204,22 +202,33 @@ class TestContext {
   }) async {
     // TODO(https://github.com/dart-lang/webdev/issues/1591): Support compiling
     // with sound null-safety in Frontend Server.
-    if (compilationMode == CompilationMode.frontendServer &&
-        nullSafety == NullSafety.sound) {
-      throw Exception(
-          'Frontend Server compilation does not support sound null-safety. See https://github.com/dart-lang/webdev/issues/1591');
-    }
+    // if (compilationMode == CompilationMode.frontendServer &&
+    //     nullSafety == NullSafety.sound) {
+    //   throw Exception(
+    //       'Frontend Server compilation does not support sound null-safety. See https://github.com/dart-lang/webdev/issues/1591');
+    // }
 
-    sdkConfigurationProvider ??= DefaultSdkConfigurationProvider();
+    // Generate SDK assets if needed.
+    // Frontend server:
+    //  - need to generate SDK js (normally included in flutter SDK).
+    //  - need to generate SDK summary for weak null safety mode as it is not
+    //    provided by the SDK installation.
+    // Build daemon:
+    //  - need to generate the weak sdk summary before the build can generate
+    //    SDK js.
+    // TODO(annagrin): update to only generate SDK JavaScript for frontend
+    // server after we have no uses of weak null safety.
+    final sdkConfigurationProvider =
+        (compilationMode == CompilationMode.frontendServer ||
+                nullSafety == NullSafety.weak)
+            ? GeneratingSdkConfigurationProvider(verboseCompiler)
+            : DefaultSdkConfigurationProvider();
 
-    if (!_sdkAssetsGenerated) {
-      _sdkAssetsGenerated = true;
-      final configuration = await sdkConfigurationProvider.configuration;
-      await generateSdkAssets(
-        configuration: configuration,
-        soundNullSafety: nullSafety == NullSafety.sound,
-        verbose: verboseCompiler,
-      );
+    final configuration = await sdkConfigurationProvider.configuration;
+    if (nullSafety == NullSafety.sound) {
+      configuration.validateSoundSummaries();
+    } else {
+      configuration.validateWeakSummaries();
     }
 
     try {